cambio de atributo

+++ domain/model/src/main/java/radar_de_aprendiz/model/aprendiz/Aprendiz.java
@@ -14,5 +14,5 @@ import java.util.List;
 @Builder(toBuilder = true)
 public class Aprendiz {
     private String nombre;
-    private List<Integer> calificaciones;
+    private List<Double> calificaciones;
 }
